sstables: take generic mutation_reader instead of memtable
The current sstable write interface only knows how to write a memtable.
For compaction, we also want it to be able to write the compaction's
output, which we can represent as a mutation_reader. So this patch
changes the sstable::write_components() method to accept a mutation_reader,
and whatever else is needed (a schema and the number of partitions in
the reader - or an estimate thereof).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>