API: Add helper function to the API
This adds a general map_to_key_value method that can work with any map
like object. This method gets the target vector in a parameter to help
the compiler inference the parameters type.

map_keys is a helper function that gets a map like objects and returns a
vector with its keys (i.e. the call to it's iterator first)

split is a wraper around the boost split and split_cf is a specific
split that is aligned with how column family are passed in the API
parameters.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>