sstable: method for estimating number of partitions in sstable
The sstable has a lot of data, but suprisingly, and accurate count of the
number of partitions isn't available. We can get a good estimate by looking
at the number of summary entries.

Based on Origin's IndexSummary.getEstimatedKeyCount().

We need this estimate for compaction if we can't get (yet) a better
estimate from the cardinality estimator algorithm.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>