database: futurize seal_active_memtables
By doing this, it is possible to synchronously wait for the seal to complete by
waiting on this future. This is useful in situations where we want to
synchronously flush data to disk.

Existing callers will not be patched, and this keeps their current behavior,
alas, asynchronously initiating a write, is preserved.

TODO: A better interface would guarantee that all writes before this one are
also complete

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>