pipe: avoid needless use of exceptions
When closing the read-end side of the pipe, we need to notify the writer
only if it was blocked in write() on a full pipe - because if it wasn't
blocked, it will get the broken pipe on the next write() attempt.

We used queue::abort() for that, which is fine, because this function
won't do anything if nobody is waiting on a full buffer. But we still
call std::make_exception_ptr() unconditionally, which is slow (involves
throwing an exception and catching it) and an annoying false-alarm when
trying to debug with gdb's "catch throw" (which stops on any throw).

So this patch does

       if (_buf.full()) {
            _buf.abort(std::make_exception_ptr(broken_pipe_exception()));
       }

So that in the typical case when the buffer was not full (hopefully it
is empty), we don't do anything.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>