Adding a void_json object to the json_elements
There are cases where a handler that returns a json element needs to
return a successfull empty response. This is common in async operation.
Though it is legit to return an empty string in this situation, it is
cleaner to return an empty response.

This adds a json_void class, that a method that needs to return a value
(i.e. the json_function) can return and the formatter will set the
response to empty.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>