reactor: circumvent ASAN false positive
Our debugging builds always produce the annoying false warning like:

==1199==ERROR: AddressSanitizer: alloc-dealloc-mismatch (malloc vs operator delete) on 0x62300003d500

This error could be silenced by setting ASAN_OPTIONS=alloc_dealloc_mismatch=0
(and we do this in test.py), but it's still annoying for manual runs, and
can mask real bugs of this type.

The problem appears to be that the address sanitizer doesn't understand
our new(with_align(..)) trick we used to allocate the reactor object,
and when we eventually use ordinary "delete" on this object, it thinks
we mixed malloc and delete.

This patch uses posix_memalign() and free() to allocate and free the
reactor structure, and ASAN stops complaining.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>