schema: allow for the setting of gc grace period
We have been leaving all the setter functions in the schema_builder, which is
usually fine.

The system tables, however, are not built this way, which means we need set
properties needed by them in some other way.

Most properties are right there in the constructor, but I argue that this one
does not belong there.

Instead, provide a set method so that the system tables which need it can call.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>