API: generalize the sum helper functions and add histogram support
This patch generalizd the sum helper function to accept any field as
long as it support the + operator and that it can be parrsed as json.

It adds a sum function to sum histograms it does so by:
adding the totatl, adding the sum, set the min and max
setting the avrage and variance and combining the samples.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>