Fix a race condition in node state for bootstrapped nodes
Booting a bootstrapped node had a race condition between setting and
advertising the state as boot strapped (call to
storage_service::bootstrap) and setting and advertising the
state as normal (call to storage_service::set_tokens) - as such a node
could get into a state in which it was "stuck" in bootstrap mode.

Following this patch you must wait for 5 seconds to have the cluster in
a stable state.

Signed-off-by: Shlomi Livne <shlomi@cloudius-systems.com>