cql3: make sure enough rows is returned in queries with IN clause
In case of queries with IN on partition key, ORDER BY and LIMIT it is
not known until after post-query sort which rows should be included in
the result set. To make sure that the output is correct each partition
specified in IN clause is queried for LIMIT rows and the excess data is
trimmed after the results are sorted.

Signed-off-by: Paweł Dziepak <pdziepak@cloudius-systems.com>