transport/server: Fail gracefully on unknown change and target type
If the switch statements don't match to change or target type
enumerations, throw a invalid_argument exception that's translated to
proper SERVER_ERROR. The error is not triggerable in practice as
change_type and target_type are enumerations. However, the compiler is
unhappy if we don't have them so lets just clean them up.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>