Cleaning the estimated_histogram
This do the following chagnes in the estimated_histogram, it uses
int64_t over unsigned to be compatible to origin and the API.

It adds a getter to the buckets and change the getteer to the
bucket_offset to be const.

It adds a get min and max similiar to origin. And it adds a merge
function to merge estimated histogram.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>