main: better error message for unsupported log level
If someone uses "--logger-log-level" with an unknown log level, e.g.,
"--logger-log-level repair=hello", the exception currently reported to the
user is:

   bad lexical cast: source type value could not be interpreted as target

In a long command line, the user has no idea where this exception is coming
from. This patches catches this exception, and instead reports a runtime_error
with the text:

   Unknown log level 'hello'

Which hopefully should be much more user friendly.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>