Fix abstract_replication_strategy's logger
The existing code tried to make the logger a function-local static object.
This attempt is commendable, because it means the logger is only created
when first used - if used at all. But it doesn't play well with our logging
infrastructure, which *assumes* that all logger objects are created during
initialization, and register themselves, to make it possible to implement
the "--logger-log-level" and "--help-loggers" command line parameters.

So the logger needs to be a global object, not a function-local object.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>