sstables: add a comment describing some sstable fields.
The reason is that the reader may think that these fields store
some statistics information about a sstable just loaded, but
they are only used when writing a new sstable.
Now I'm starting to see the value of having a sstable class for
a sstable loaded and another one for a sstable being created
(that's what Origin does).

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>