db/schema_tables: Wire up CF creation CQL event
The code is merge_tables() is a twisted maze of tricks that is hard to
restructure so that event notification can be done cleanly like with
keyspaces.

The problem there is that we need to run bunch of database operations
for the merging that really need to happen on all the shards.  To fix
the issue, lets cheat a little and simply only run CQL event
notification on cpu zero.

This seems to fix cluster schema propagation issues in urchin-dtest. I
can now run TestSimpleCluster.simple_create_insert_select_test without
any additional delays inserted into the test code.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>